Working MPC Controller
Defining a cost function and solving it using CVX, an optimal control input is determined using an MPC problem formulation. The following issues still need to be addressed: (1) prediction horizon (when the sample time is too small, the prediction horizon is not long enough, resulting in instability) (2) reference tracking needs to be checked to see if it works. possible dc gain issues could occur (3) simulation with large time steps are very jittery, which could be improved with smoothing inter-sample interpolation