Implemented time-dependent radius and rotational velocity with adaptive control framework
LQR control is used and a new EP is obtained based on the desired Radius and Omega at that time step k. The DARE function is solved each iteration to determine a new control gain on the updated linearized system matrices. This framework allows for easy portability to the MPC paradigm